refactor(api): extract hooks + commands handlers from system.rs (#3749 4/N)
Continue the routes/system.rs split started in #4452 (pairing),
webhook-trigger and chat-command-catalog endpoints
(`webhook_wake`, `webhook_agent`, `list_commands`, `get_command`,
plus the constant-time `validate_webhook_token` helper) into a new
sibling module `routes/hooks_commands.rs`.

`system::router()` now mounts these via
`.merge(crate::routes::hooks_commands::router())` -- public route
paths (`/api/hooks/wake`, `/api/hooks/agent`, `/api/commands`,
`/api/commands/{name}`) are unchanged. The glob re-export in
`routes/mod.rs` keeps `routes::webhook_wake` / `routes::list_commands`
resolvable so existing utoipa bindings in `openapi.rs` and the route
registration in `server.rs` keep compiling without edits.

system.rs goes from 5939 -> 5638 lines (-301); the new
hooks_commands.rs is 328 lines.

Refs #3749 (4 of N sub-domain extractions).
